# service/docs/bruno/openapi/membership-management.yaml
openapi: 3.1.0
info:
  title: Membership Management API
  description: API for managing memberships, including status checks, cancellations, monthly payments, and activation.
  version: 1.2.0
servers:
  - url: https://api.membership.com
    description: Production server
  - url: http://localhost:3000
    description: Local development server
paths:
  /api/memberships/{userId}/status:
    get:
      summary: Check membership status
      operationId: checkMembershipStatus
      description: Checks if the specified user has an active membership.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            example1: "12345"
          description: The unique ID of the user.
      responses:
        '200':
          description: Membership status retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasMembership:
                    type: boolean
                    description: Indicates if the user has an active membership.
                    example2: true
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: Error code.
          example6: 400
        message:
          type: string
          description: Error message.
          example7: "Invalid user ID."
      required:
        - code
        - message
...